Use checkboxes for binary display options
The search form "display options" section includes three different options which offer a binary choice -- AS dates, glossary links, and armory descriptions.

Rather than showing those as a pull-down select menu with two choices, this change causes them to be presented as a checkbox, which better communicates the range of options.