Allow negative weights for complex search
This change enables you to enter negative integers in the weight fields the complex search. Weights are summed as usual, and items with a score of less than one are excluded from the result set.

For example, you can find armory with a red bear that is not fieldless by searching for description "BEAR:gules" with weight 1 and description "NO" with weight -1.