Add minimum-score field to complex search form
This change adds a field to the display options on the complex search form for the minimum score to returned by the database server.

It defaults to 1, corresponding to the current behavior of returning any items that match at least one search criteria.

You can set it to a larger integer to only see items which match multiple rules, according to the standard weight logic.

A number less than one in the minimum-score field will be subtracted from the maximum possible weight from all of your search lines.

For example, setting the minimum to -1 will only return rows with the maximum possible score or one point lower than that, corresponding to typical use of the complex search form for armory conflict-checking.

Setting the minimum to 0 will only return rows with the maximum possible score.

Internally this uses the database server's existing "t" threshold command.